pyproject.toml
[project]
name = "geotab-fleet-mcp"
version = "0.1.0"
description = "MCP server for conversational fleet management via Geotab MyGeotab API"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "mygeotab>=0.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "duckdb>=1.0.0",
    "pydantic>=2.0.0",
    "flask>=3.0.0",
    "google-genai>=1.0.0",
]

[project.scripts]
geotab-mcp = "geotab_mcp.server:main"
geotab-dashboard = "geotab_mcp.dashboard:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/geotab_mcp"]
